JFDS 939
Regarding the OP, Please remove scalloped when we need to lipp the OP, but this should remain for doors. The OP core width should be the same as the door leaf, the system is reducing OP core width. Also why is the door height reduced as in this example there is no adjustment on door leaf height?