JFDS-1001
For the LegX2 columns on all cut list can we update please when 4 sided frame is selected? The formula currently in system is O/A Frame height -frame thickness X1.    Please update when a 4 sided frame is selected to O/A Frame height -frame thickness X2.            NB: THIS IS ONLY FOR 4 SIDED FRAME